+++ packages/server/test/unit/screenshots_spec.js
@@ -744,7 +744,7 @@ describe('lib/screenshots', () => {
         scaled: true,
         blackout: [],
         startTime: '2018-06-27T20:17:19.537Z',
-        specName: 'integration/spec.coffee',
+        specName: 'integration/spec.cy.ts',
       }
 
       this.details = {
@@ -754,7 +754,7 @@ describe('lib/screenshots', () => {
         multipart: false,
         pixelRatio: 1,
         name: 'my-screenshot',
-        specName: 'integration/spec.coffee',
+        specName: 'integration/spec.cy.ts',
         testFailure: true,
         path: '/path/to/my-screenshot.png',
       }
@@ -775,7 +775,7 @@ describe('lib/screenshots', () => {
           multipart: false,
           pixelRatio: 1,
           name: 'my-screenshot',
-          specName: 'integration/spec.coffee',
+          specName: 'integration/spec.cy.ts',
           testFailure: true,
           path: '/path/to/my-screenshot.png',
           scaled: true,
@@ -804,7 +804,7 @@ describe('lib/screenshots', () => {
           multipart: false,
           pixelRatio: 1,
           name: 'my-screenshot',
-          specName: 'integration/spec.coffee',
+          specName: 'integration/spec.cy.ts',
           testFailure: true,
           path: '/new/path/to/screenshot.png',
           scaled: true,
@@ -826,7 +826,7 @@ describe('lib/screenshots', () => {
           multipart: false,
           pixelRatio: 1,
           name: 'my-screenshot',
-          specName: 'integration/spec.coffee',
+          specName: 'integration/spec.cy.ts',
           testFailure: true,
           path: '/path/to/my-screenshot.png',
           scaled: true,