+++ system-tests/projects/e2e/cypress.config.js
@@ -5,7 +5,7 @@ module.exports = defineConfig({
   allowCypressEnv: false,
   retries: null,
   e2e: {
-    specPattern: 'cypress/e2e/**/*.{js,jsx,mjs,ts,tsx,coffee}',
+    specPattern: 'cypress/e2e/**/*.{js,jsx,mjs,ts,tsx}',
     setupNodeEvents (on, config) {
       return plugin(on, config)
     },