+++ system-tests/projects/ids/cypress/e2e/foo.js
@@ -1,0 +1,7 @@
+describe('foo [000]', () => {
+  it('bars [001]', () => {
+    expect('foo.js').to.equal('foo.js')
+  })
+
+  it('quux [002]')
+})