+++ system-tests/projects/react-webpack-interop/cypress.config.js
@@ -4,7 +4,7 @@ module.exports = defineConfig({
   allowCypressEnv: false,
   e2e: {
     supportFile: false,
-    specPattern: 'cypress/e2e/**/*.{js,jsx,mjs,ts,tsx,coffee}',
+    specPattern: 'cypress/e2e/**/*.{js,jsx,mjs,ts,tsx}',
     setupNodeEvents (on, config) {},
   },
 })