+++ system-tests/projects/todos/cypress.config.js
@@ -15,7 +15,7 @@ module.exports = {
   },
   'e2e': {
     'supportFile': 'tests/_support/spec_helper.js',
-    'specPattern': 'tests/**/*.(js|ts|coffee)',
+    'specPattern': 'tests/**/*.(js|ts)',
     'setupNodeEvents': (on, config) => config,
   },
 }