+++ system-tests/test/async_timeouts_spec.js
@@ -7,5 +7,8 @@ describe('e2e async timeouts', () => {
     spec: 'async_timeouts.cy.js',
     snapshot: true,
     expectedExitCode: 2,
+    config: {
+      screenshotOnRunFailure: false,
+    },
   })
 })